add account api

+++ src/main/java/ru/sirius/natayarik/ft/data/AccountDTO.java
@@ -6,24 +6,24 @@ package ru.sirius.natayarik.ft.data;
 
 public class AccountDTO {
 
-    private int id;
-    private int userId;
+    private long id;
+    private long userId;
     private String name;
     private CurrencyDTO currency;
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
-    public int getUserId() {
+    public long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(long userId) {
         this.userId = userId;
     }
